pyproject.toml
[project]
name = "automated-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-cov (>=7.0.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    # Ignore the tests directory
    "tests/*",
    # Ignore virtual environments
    "venv/*",
    # Ignore a specific file
    "utils/tirefire.py",
    # Ignore all __init__.py files
    "*/__init__.py",
]